model/requirements.txt
kafka-python==1.4.3
requests==2.18.4
